refactor: move skeleton to ui
